Add typed state and event handlers
Uniflow exposes some extension functions for observing states and events in your UI. `onStates` & `onEvents` work great but only they provide open types. This is useful when you are dealing with states and events of multiple types. For example:

```kotlin
onStates(viewModel) {
    when (it) {
        is UIState.Loading -> handleLoading()
        is MyCustomExampleState.Success -> handleSuccess(it.stuff)
    }
}
```
In this example `it` is `UIState`.
But the majority of the time we will actually only use our custom states and events like this.

```kotlin
onStates(viewModel) {
    when (it) {
        is MyCustomExampleState.Loading -> handleLoading()
        is MyCustomExampleState.Success -> handleSuccess(it.stuff)
    }
}
```
`MyCustomExampleState` would typically be a sealed class and we would want to handle it exhaustively but there are no warning because it is `UIState` which is an open class.

These extensions will provide was with strongly type state and event handlers which means you will be warned if you don't handle all permutations of the sealed class.

```kotlin
onStatesOfType<SampleLegacyDBState>(viewModel) {
    when (it) {
        is MyCustomExampleState.Loading -> handleLoading()
        is MyCustomExampleState.Success -> handleSuccess(it.stuff)
    }
}
```